RR ahould be done

+++ scheduler.c
@@ -271,7 +271,9 @@ typedef struct PCB {
 int* Process_Wait;
 float* WTA;
 
-PriorityQueue* readyQueue;
+PriorityQueue * readyQueue;
+CircularQueue * cq;
+
 PCB* currentPCB = NULL;// Currently running process
 int Start_execution=0;
 FILE* log_file; 
@@ -304,7 +306,9 @@ void checkWaitingQueue() {
                 insertRuntimePriorityQueue(readyQueue, waiting_pcb->process);
             } else if (algo == 2) {
                 insertPriorityPriorityQueue(readyQueue, waiting_pcb->process);
-            } // ... handle other algorithms
+            } else if (algo == 3) {
+                enqueueCircularQueue (cq, waiting_pcb->process);
+            }
         } else {
             // If still can't allocate, put back in waiting queue
             enqueueWaiting(memory_waiting_queue, waiting_pcb);
@@ -446,7 +450,9 @@ int main(int argc, char* argv[])
     }
     if(algo==1)
     {
-        PCB pcb [process_count];
+        PCB* pcb [process_count];
+        PCB* insertedPCB = NULL;
+
         int total_processes = 0; 
         readyQueue = createPriorityQueue(process_count);
         struct msgbuff1
@@ -472,40 +478,23 @@ int main(int argc, char* argv[])
                     .waiting_time = 0,
                     .pid = -1  // Not yet forked
                 };
-                pcb [i] = newPCB;
-                insertRuntimePriorityQueue(readyQueue, pcb[i].process);
+                pcb [i] = &newPCB;
+                insertRuntimePriorityQueue(readyQueue, pcb[i]->process);
+                insertedPCB = pcb[i];
                 processCount++;
                 i++;
             }
 
             // If no process is running, select the next process
-            if (currentPCB == NULL && readyQueue->size > 0) {
+            if (currentPCB == NULL && readyQueue->size > 0) 
+            {
                 Process nextProcess = removeRuntimePriorityQueue(readyQueue);
+                scheduleNextProcess(insertedPCB->process, pcb);
 
-                // Fork and execute the process
-                pid_t pid = fork();
-                if (pid == 0) {
-                    // Child process simulates execution
-                    char runtimeChar[20];
-                    sprintf(runtimeChar, "%d", nextProcess.runtime);
-                    execl("./process.out", "./process.out", runtimeChar, (char*)NULL);
-                    perror("Failed to execute process");
-                    exit(1);
-                } else if (pid > 0) {
-                    // Parent process tracks the process
-                    currentPCB = (PCB*)malloc(sizeof(PCB));
-                    currentPCB->process = nextProcess;
-                    currentPCB->pid = pid;
-                    currentPCB->state = RUNNING;
-                    currentPCB->start_time = getClk();
-                    currentPCB->remaining_time = nextProcess.runtime;
-                    printf("usdhcusdhuvs\n\n\n");
-                    logProcessEvent(log_file, currentPCB, "started", currentPCB->start_time);
-                } else {
-                    perror("Fork failed");
-                    exit(1);
-                }
+
+                
             }
+
 
             // Exit condition: no running process and no more processes in the queue
             if(!currentPCB && (lasttime+1) == getClk())
@@ -521,7 +510,6 @@ int main(int argc, char* argv[])
     else if (algo == 2) 
     {
     readyQueue = createPriorityQueue(process_count);
-        int lastime=0;
         bool recieved=false;
         PCB* pcb[process_count];
         for (int i = 0; i < process_count; i++) {
@@ -557,21 +545,25 @@ int main(int argc, char* argv[])
 
         if (currentPCB) {
             // Preemption logic
-            if (insertedPCB && currentPCB->process.priority > insertedPCB->process.priority) {
+            if (insertedPCB && currentPCB->process.priority > insertedPCB->process.priority && currentPCB->memory_block->is_free) 
+            {
                 printf("Preempting process %d (pid=%d)\n", currentPCB->process.id, currentPCB->pid);
                 kill(currentPCB->pid, SIGTSTP);  // Stop current process
                  int i = 0;
                 bool found = false;
-                while (i < process_count && !found) {
+                while (i < process_count && !found) 
+                {
                     if (pcb[i] && pcb[i]->process.id == currentPCB->process.id)
-                    {int elapsed_time=getClk()-Start_execution;
-                     pcb[i]->remaining_time-=elapsed_time;
-                     Start_execution=0;
-                       pcb[i]->Last_execution=getClk(); 
+                    {
+                        int elapsed_time=getClk()-Start_execution;
+                        pcb[i]->remaining_time-=elapsed_time;
+                        Start_execution=0;
+                        pcb[i]->Last_execution=getClk(); 
                     }
                     
                     i++;
                 }
+
                 currentPCB->process.prempted = true;
                 currentPCB->state = Blocked;
                 logProcessEvent(log_file, currentPCB, "Blocked", getClk());
@@ -585,13 +577,14 @@ int main(int argc, char* argv[])
             }
             else
             {
-                 if (insertedPCB) {
-                printf("I am inserted process %d",insertedPCB->process.id);
-                insertPriorityPriorityQueue(readyQueue, insertedPCB->process);
-                free(insertedPCB);
-                insertedPCB = NULL;
-                recieved=true;
-            }
+                 if (insertedPCB) 
+                {
+                    printf("I am inserted process %d\n",insertedPCB->process.id);
+                    insertPriorityPriorityQueue(readyQueue, insertedPCB->process);
+                    free(insertedPCB);
+                    insertedPCB = NULL;
+                    recieved=true;
+                }
 
             }
 
@@ -600,7 +593,7 @@ int main(int argc, char* argv[])
         else {
             // If no current process, add new process to the queue
             if (insertedPCB) {
-                printf("I am inserted process %d",insertedPCB->process.id);
+                printf("I am inserted process %d\n",insertedPCB->process.id);
                 insertPriorityPriorityQueue(readyQueue, insertedPCB->process);
                 free(insertedPCB);
                 insertedPCB = NULL;
@@ -636,10 +629,10 @@ int main(int argc, char* argv[])
         // Exit condition: no running process and no more processes in the queue
        if(finished_counter==process_count)
        break;
-        if(!currentPCB&&!insertedPCB&&(lastime+1)==getClk())
+        if(!currentPCB&&!insertedPCB&&(lasttime+1)==getClk())
         {Waiting++;
             printf("Waiting= %d \n",Waiting);
-            lastime=getClk();
+            lasttime=getClk();
         
         }
         
@@ -648,7 +641,7 @@ int main(int argc, char* argv[])
 
 else if (algo == 3)
 {
-    CircularQueue *cq =  createCircularQueue (process_count);   // ready queue
+    cq =  createCircularQueue (process_count);   // ready queue
 
     log_file = fopen("scheduler.log", "w");
     perf_file = fopen("scheduler.perf", "w");
@@ -682,7 +675,16 @@ else if (algo == 3)
             new_pcb->pid = -1; 
 
             pcb [total_processes] = *new_pcb;
-            enqueueCircularQueue (cq, pcb [total_processes].process);
+
+            // allocate memory for new processes if size is available, otherwise add to waiting list
+            MemoryBlock * newBlock = allocateMemory (memory_allocator, new_pcb->process.memsize, new_pcb->process.id);
+            if (newBlock != NULL)
+            {
+                new_pcb->memory_block = newBlock;
+                enqueueCircularQueue (cq, pcb [total_processes].process);
+            }
+            else enqueueWaiting(memory_waiting_queue, new_pcb);
+            
             total_processes++;
         }   
         
@@ -702,7 +704,6 @@ else if (algo == 3)
         // Select process to run
         Process currentProcess = dequeueCircularQueue (cq);
         // Find PCB for this process 
-        //struct PCB *current_process = NULL; 
         for (int i = 0; i < total_processes; i++) 
         {
             if (pcb[i].process.id == currentProcess.id) 
@@ -721,10 +722,8 @@ else if (algo == 3)
             if (current_process->pid == 0)
             {
                 char remaining_time_str[10]; 
-                //char quantum_str[10]; 
 
                 sprintf(remaining_time_str, "%d", current_process->remaining_time); 
-                //sprintf(quantum_str, "%d", quanta); 
 
                 execl("./process.out", "./process.out", remaining_time_str, argv[1], argv[3], NULL); 
                 perror("execl failed"); 
@@ -769,6 +768,8 @@ else if (algo == 3)
             current_process->state = 2;     //finished
 
             write_log (current_process, "finished", current_time);
+            freeMemory (memory_allocator, current_process->memory_block, current_process->process.id);
+            checkWaitingQueue();
 
             int status;
             waitpid (current_process->pid, &status, 0);
@@ -969,6 +970,8 @@ else if (algo == 4)
 
 // Handle child process termination
 void handleProcessCompletion(int signum) {
+    if (algo < 3)
+    {
     int status;
     pid_t pid = waitpid(-1, &status, WNOHANG);
     if (pid > 0 && currentPCB != NULL && currentPCB->pid == pid) {
@@ -993,6 +996,7 @@ void handleProcessCompletion(int signum) {
             free(currentPCB);
         currentPCB = NULL;
         kill(getppid(),SIGUSR1);
+    }
     }
 }
